requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1.5
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
ecdsa==0.14.1
gunicorn==20.0.4
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
pyasn1==0.4.8
PyJWT==2.0.0
python-jose==3.2.0
pytz==2020.5
requests==2.25.1
rsa==4.6
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.26.2
whitenoise==5.2.0
